backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
sse-starlette==2.1.0
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30

# Cache / Queue
redis[hiredis]==5.0.6
arq==0.25.0

# Settings
pydantic==2.7.1
pydantic-settings==2.3.0

# LLM providers
groq==0.9.0
openai==1.30.1               # embeddings only
anthropic==0.28.0
tiktoken==0.7.0

# Document extraction
pypdfium2==4.30.0
pdfplumber==0.11.1
pytesseract==0.3.10
python-docx==1.1.2
Pillow==10.3.0
trafilatura==1.9.0
httpx==0.27.0
pandas==2.2.2

# ML / embeddings
sentence-transformers==3.0.1
numpy==1.26.4

# Observability
opentelemetry-sdk==1.24.0
opentelemetry-api==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
prometheus-client==0.20.0

# MLflow
mlflow==2.13.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bleach==6.1.0
detect-secrets==1.5.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0

# Code quality
ruff==0.4.7
mypy==1.10.0

# Misc
structlog==24.2.0
APScheduler==3.10.4
cryptography==42.0.7
